+++ packages/react-native-the-sheet/CHANGELOG.md
@@ -1,5 +1,19 @@
 # react-native-the-sheet
 
+## 2.0.8
+
+### Patch Changes
+
+- BottomSheetScrollViewProps has a more defined prop interface (⚠️ **BREAKING**)
+  - Remove `styles` prop
+- BottomSheetFlatListProps has a more defined prop interface (⚠️ **BREAKING**)
+  - Remove `styles` prop
+
+- Rename `SheetKeyboardProviderContextType` to `SheetKeyboardContextType` (⚠️ **BREAKING**)
+- Rename `useSheetKeyboardProvider` to `useSheetKeyboard` (⚠️ **BREAKING**)
+
+- Add `useHeightBudget`, `HeightBudgetProvider`, `HeightClaim`, and `HeightFill` to support for deeply nested dynamic sizing scroll views
+
 ## 2.0.7
 
 ### Patch Changes